[ESI][Cosim] Reducing MMIO data widths to 32 bit (#6017)
Some platforms only support 32 bit MMIO, so in accordance with the least
common denominator rule...

+++ include/circt/Dialect/ESI/cosim/dpi.h
@@ -51,12 +51,12 @@ DPI int sv2cCosimserverMMIORegister();
 
 /// Read MMIO function pair. Assumes that reads come back in the order in which
 /// they were requested.
-DPI int sv2cCosimserverMMIOReadTryGet(int *address);
-DPI void sv2cCosimserverMMIOReadRespond(long long data, char error);
+DPI int sv2cCosimserverMMIOReadTryGet(uint32_t *address);
+DPI void sv2cCosimserverMMIOReadRespond(uint32_t data, char error);
 
 /// Write MMIO function pair. Assumes that write errors come back in the order
 /// in which the writes were issued.
-DPI int sv2cCosimserverMMIOWriteTryGet(int *address, long long *data);
+DPI int sv2cCosimserverMMIOWriteTryGet(uint32_t *address, uint32_t *data);
 DPI void sv2cCosimserverMMIOWriteRespond(char error);
 
 #ifdef __cplusplus
